fix: update test precision and benchmark format for col2im
